refactor: extract File resource access into FileAccessChecker
Move isAvailable, isDirectlyAccessible and getAccessConditions out of
FileResourceType into a dedicated, EDT-independent FileAccessChecker so
the access logic can be reused when the resource type is migrated to API
Platform 3.0.
